+++ Cohort/Domain/ErasureSubjectAttribute.cs
@@ -1,4 +1,4 @@
 namespace Cohort.Domain;
 
-[AttributeUsage(AttributeTargets.Property, Inherited = false, AllowMultiple = false)]
+[AttributeUsage(AttributeTargets.Property, Inherited = false, AllowMultiple = true)]
 public sealed class ErasureSubjectAttribute : Attribute;